+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed a bug when evaluating a generative task that uses logprobs with partial
+  completions. The issue was that the first label token mapping wasn't updated
+  immediately after the dataset config changed, which has been fixed now.
+
 ## [v17.1.0] - 2026-03-24
 
 ### Added